requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fake-useragent==2.2.0
google-auth==2.49.0
idna==3.11
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
lxml==5.3.0
urllib3==2.6.3
pandas
redis

# ML / AI
transformers
torch
accelerate
pymupdf
sentence-transformers
supabase

# Google APIs (kept unpinned intentionally)
google-auth-oauthlib
google-api-python-client

# -----------------------------
# 🔴 Duplicates / old versions (commented)
# -----------------------------

# requests==2.31.0
# beautifulsoup4==4.12.2
# fake-useragent==1.5.1
# google-auth
